exe suffix, stdin option
